moved everything into udMathUtility.h
